// src/app/interfaces/usuarios.interface.ts
// Generated by https://quicktype.io

//export interface Usuarios {
    //id?:       number;
    //username: string;
    //password: string;
//}

export interface Usuarios {
    // username: string;
    // password: string;
    id_usuario?: number;
    id_estado: number;
    nombres: string;
    apellidos: string;
    email: string;
    contrasena: string;
    id_rol: number;
    numero_intento: number;
}